list users with roles

+++ app/Http/Controllers/UserController.php
@@ -44,12 +44,15 @@ class UserController extends Controller
     {
         // $user = User::findOrFail($id);
         // dd($user);
+        $user->load('roles');
         return view('users.show', compact('user'));
     }
 
     public function edit(User $user)
     {
-        return view('users.edit', compact('user'));
+        $roles = Role::all()->pluck('name', 'id');
+        $user->load('roles');
+        return view('users.edit', compact('user', 'roles'));
     }
 
     public function update(UserEditRequest $request, User $user)
@@ -69,11 +72,18 @@ class UserController extends Controller
         // }
 
         $user->update($data);
+
+        $roles = $request->input('roles', []);
+        $user->syncRoles($roles);
         return redirect()->route('users.show', $user->id)->with('success', 'Usuario actualizado correctamente');
     }
 
     public function destroy(User $user)
     {
+        if (auth()->user()->id == $user->id) {
+            return redirect()->route('users.index');
+        }
+
         $user->delete();
         return back()->with('succes', 'Usuario eliminado correctamente');
     }
